Add read/write-through cache for fastboot cookies
This adds a cache for fastboot cookies so that cookies that have just
been written can be read with the new value afterwards. This isn't possible
by default as cookies are read from the request but written to the response in
express.

This also makes sure that cookies that were written for a different path etc.
and might not be accessible for the current request are not returned from the
cache.